requirements.txt
croniter==1.3.4
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.11
requests==2.27.1
feedparser==6.0.8
python-dateutil==2.8.2
beautifulsoup4==4.12.2
